+++ packages/perseus/src/widgets/radio/__stories__/choice-none-above.stories.jsx
@@ -27,6 +27,7 @@ const ChoiceDefaults = {
     reviewMode: false,
     showRationale: false,
     showCorrectness: false,
+    multipleSelect: false,
     crossedOut: false,
     previouslyAnswered: false,
     apiOptions: {},