+++ packages/perseus/src/widgets/radio/__stories__/focus-ring.stories.jsx
@@ -1,38 +1,44 @@
 // @flow
 import * as React from "react";
 
+import * as styleConstants from "../../../styles/constants.js";
 import FocusRing from "../focus-ring.jsx";
 
-type StoryArgs = {||};
+type StoryArgs = {|
+    children: React.Node,
+    color: string,
+    visible: boolean,
+    multipleSelect: boolean,
+|};
 
 type Story = {|
     title: string,
+    args: StoryArgs,
 |};
 
 export default ({
     title: "Perseus/Widgets/Radio/Focus Ring",
+    args: {
+        children: "",
+        color: styleConstants.kaGreen,
+        visible: true,
+        multipleSelect: false,
+    },
 }: Story);
 
-export const EmptyPropsObject = (args: StoryArgs): React.Node => {
-    return <FocusRing />;
-};
+export const Interactive = (args: StoryArgs): React.Node => {
+    // faux choice is just for demonstration
+    const fauxChoiceStyles = {
+        height: "20px",
+        width: "20px",
+        background: args.color,
+        borderRadius: args.multipleSelect ? "3px" : "50%",
+    };
 
-export const VisibleSetToTrue = (args: StoryArgs): React.Node => {
-    return <FocusRing visible={true} />;
-};
+    const customArgs = {
+        ...args,
+        children: args.children || <div style={fauxChoiceStyles} />,
+    };
 
-export const VisibleSetToFalse = (args: StoryArgs): React.Node => {
-    return <FocusRing visible={false} />;
-};
-
-export const ColorRed = (args: StoryArgs): React.Node => {
-    return <FocusRing color="red" />;
-};
-
-export const SingleSpanChild = (args: StoryArgs): React.Node => {
-    return (
-        <FocusRing>
-            <span>test</span>
-        </FocusRing>
-    );
+    return <FocusRing {...customArgs} />;
 };