+++ packages/perseus/src/widgets/radio/focus-ring.jsx
@@ -8,40 +8,47 @@ import * as React from "react";
 
 import * as styleConstants from "../../styles/constants.js";
 
-class FocusRing extends React.Component<{
-    children?: any,
+type Props = {|
+    children?: React.Node,
     // Whether the focus ring is visible. Allows for positioning
     // the child identically regardless of whether the ring is visible.
     visible: boolean,
     // Color of the focus ring
     color: string,
-    ...
-}> {
-    static defaultProps: {|color: string, visible: boolean|} = {
-        visible: true,
-        color: styleConstants.kaGreen,
-    };
+    // Whether a user can select multiple options or not
+    multipleSelect: boolean,
+|};
 
-    render(): React.Element<"span"> {
-        const borderColor = this.props.visible
-            ? this.props.color
-            : "transparent";
-        const style = {
-            borderColor: borderColor,
-        };
-        return (
-            <span className={css(styles.ring)} style={style}>
-                {this.props.children}
-            </span>
-        );
-    }
+function FocusRing(props: Props): React.Node {
+    const {visible, color, children, multipleSelect} = props;
+
+    const borderColor = visible ? color : "transparent";
+    const borderRadius = multipleSelect ? 5 : "50%";
+    const style = {
+        borderColor,
+        borderRadius,
+    };
+    return (
+        <span
+            data-test-id="focus-ring"
+            className={css(styles.ring)}
+            style={style}
+        >
+            {children}
+        </span>
+    );
 }
+
+FocusRing.defaultProps = {
+    visible: true,
+    color: styleConstants.kaGreen,
+    multipleSelect: false,
+};
 
 const styles = StyleSheet.create({
     ring: {
         margin: "auto",
         display: "inline-block",
-        borderRadius: "50%",
         borderWidth: 2,
         padding: 2,
         borderStyle: "solid",